pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-dreaming"
version = "2.0.0"
description = "Knowledge operating system for humans and AI"
requires-python = ">=3.9"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
brain = "brain.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["brain*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
